Move 3d interface to the annotation directory
