+++ frontend/src/components/annotation/3d_annotation/AnnotationScene3D.ts
@@ -2,12 +2,12 @@ import * as THREE from 'three';
 import {OrbitControls} from './OrbitControls';
 import ObjectTransformControls from "./ObjectTransformControls";
 import CuboidDrawerTool from "./CuboidDrawerTool";
-import {Instance3D} from '../vue_canvas/instances/Instance3D';
-import {Instance} from '../vue_canvas/instances/Instance';
-import Cuboid3DInstance from "../vue_canvas/instances/Cuboid3DInstance";
+import {Instance3D} from '../../vue_canvas/instances/Instance3D';
+import {Instance} from '../../vue_canvas/instances/Instance';
+import Cuboid3DInstance from "../../vue_canvas/instances/Cuboid3DInstance";
 import {getCenterPoint} from "./utils_3d";
 import Vue from 'vue';
-import {LabelFile} from "../../types/label";
+import {LabelFile} from "../../../types/label";
 
 export default class AnnotationScene3D {
   public scene: THREE.Scene;
