+++ frontend/src/components/annotation/3d_annotation/ObjectTransformControls.ts
