+++ frontend/src/components/annotation/3d_annotation/canvas_3d.vue
