+++ frontend/src/components/annotation/3d_annotation/hotkeys_sensor_fusion.vue
