+++ frontend/src/components/annotation/3d_annotation/sensor_fusion_editor.vue
@@ -239,22 +239,21 @@
     has_duplicate_instances,
     add_ids_to_new_instances_and_delete_old,
     check_if_pending_created_instance
-  } from '../annotation/image_annotation/utils/AnnotationSavePrechecks';
+  } from '../image_annotation/utils/AnnotationSavePrechecks';
   import toolbar_sensor_fusion from "./toolbar_sensor_fusion";
-  import instance_detail_list_view from "../annotation/image_annotation/instance_detail_list_view";
+  import instance_detail_list_view from "../image_annotation/instance_detail_list_view";
   import context_menu_3d_editor from "./context_menu_3d_editor";
   import canvas_3d from "./canvas_3d";
-  import main_menu from '../main_menu/menu'
-  import error_multiple from '../regular/error_multiple'
+  import main_menu from '../../main_menu/menu'
+  import error_multiple from '../../regular/error_multiple'
   import moment from "moment";
-  import axios from "../../services/customInstance";
-  import * as instanceServices from '../../services/instanceServices';
+  import axios from "../../../services/customInstance";
+  import * as instanceServices from '../../../services/instanceServices';
   import FileLoader3DPointClouds from "./FileLoader3DPointClouds";
-  import * as instance_utils from "../../utils/instance_utils.ts"
-  import * as THREE from "three";
-  import {UpdateInstanceCommand} from "../annotation/image_annotation/commands/update_instance_command.ts";
-  import {CommandManagerAnnotationCore} from "../annotation/image_annotation/annotation_core_command_manager";
-  import {trackTimeTask, finishTaskAnnotation} from "../../services/tasksServices";
+  import * as instance_utils from "../../../utils/instance_utils.ts"
+  import {UpdateInstanceCommand} from "../image_annotation/commands/update_instance_command.ts";
+  import {CommandManagerAnnotationCore} from "../image_annotation/annotation_core_command_manager";
+  import {trackTimeTask, finishTaskAnnotation} from "../../../services/tasksServices";
 
   export default Vue.extend({
     name: "sensor_fusion_editor_3d",
