+++ frontend/src/components/annotation/3d_annotation/utils_3d.ts
