+++ frontend/src/components/vue_canvas/instances/Instance.ts
@@ -1,6 +1,6 @@
 import * as THREE from 'three';
-import AnnotationScene3D from "../../3d_annotation/AnnotationScene3D";
-import {getCenterPoint} from "../../3d_annotation/utils_3d";
+import AnnotationScene3D from "../../annotation/3d_annotation/AnnotationScene3D";
+import {getCenterPoint} from "../../annotation/3d_annotation/utils_3d";
 import {LabelColourMap} from "../../../types/label_colour_map";
 import {LabelFile} from "../../../types/label";
 import {MousePosition, Point} from "../../../types/mouse_position";
